test: AuthorizationExtractor 테스트 추가
